third_party/blink/web_tests/fast/css/padding-start-end-expected.txt

Tests padding-start and padding-end

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testWidth("ltr", "padding-inline-start", "10px") is 110
PASS testWidth("ltr", "padding-inline-end", "20px") is 120
PASS test("ltr", "padding-inline-start", "10px", "padding-left") is "10px"
PASS test("ltr", "padding-inline-end", "20px", "padding-right") is "20px"
PASS test("ltr", "padding-left", "10px", "padding-inline-start") is "10px"
PASS test("ltr", "padding-right", "20px", "padding-inline-end") is "20px"
PASS testWidth("rtl", "padding-inline-start", "10px") is 110
PASS testWidth("rtl", "padding-inline-end", "20px") is 120
PASS test("rtl", "padding-inline-start", "10px", "padding-right") is "10px"
PASS test("rtl", "padding-inline-end", "20px", "padding-left") is "20px"
PASS test("rtl", "padding-right", "10px", "padding-inline-start") is "10px"
PASS test("rtl", "padding-left", "20px", "padding-inline-end") is "20px"
PASS successfullyParsed is true

TEST COMPLETE
